feat(App.tsx, GolCanvas.tsx): Move game logic to App component instead of the Canvas component
